FeedForwardNN/mnist_recognition/requirements.txt
tensorflow-macos>=2.12.0
tensorflow-metal>=0.7.0
numpy>=1.23.0
pandas>=2.0.0
streamlit>=1.31.0
streamlit-drawable-canvas>=0.9.0
scikit-learn>=1.3.0
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.13.0
opencv-python>=4.8.0
yfinance>=0.2.0
